// projects/agmx/src/lib/services/google-maps-types.ts
export interface MVCObject {
  addListener(eventName: string, handler: Function): MapsEventListener;
}

export interface MVCArray<T> extends MVCObject {
  clear(): void;
  getArray(): Array<T>;
  getAt(i: number): T;
  getLength(): number;
  insertAt(i: number, elem: T): void;
  pop(): T;
  push(elem: T): number;
  removeAt(i: number): T;
  setAt(i: number, elem: T): void;
  /* tslint:disable */
  /*
  * Tslint configuration check-parameters will prompt errors for these lines of code.
  * https://palantir.github.io/tslint/rules/no-unused-variable/
  */
  forEach(callback: (elem: T, i: number) => void): void;
  /* tslint:enable */
}

/**
 * Custom type
 */
export interface GoogleMap<E extends Element = Element> extends google.maps.MVCObject {
  data?: google.maps.Data;
  controls: google.maps.MVCArray<Node>[];
  mapTypes: google.maps.MapTypeRegistry;
  overlayMapTypes: google.maps.MVCArray<google.maps.MapType>;

  constructor(el: HTMLElement, opts?: MapOptions): void;
  panTo(latLng: google.maps.LatLng | google.maps.LatLngLiteral): void;
  panBy(x: number, y: number): void;
  setZoom(zoom: number): void;
  getCenter(): google.maps.LatLng;
  setCenter(latLng: google.maps.LatLng | google.maps.LatLngLiteral): void;
  getBounds(): google.maps.LatLngBounds;
  getMapTypeId(): google.maps.MapTypeId;
  getZoom(): number;
  setOptions(options: MapOptions): void;
  panToBounds(latLngBounds: google.maps.LatLngBounds | google.maps.LatLngBoundsLiteral, padding?: number | google.maps.Padding): void;
  fitBounds(bounds: google.maps.LatLngBounds | google.maps.LatLngBoundsLiteral, padding?: number | google.maps.Padding): void;
  getClickableIcons(): boolean;
  getDiv(): E;
  getHeading(): number;
  getProjection(): google.maps.Projection | null;
  getStreetView(): google.maps.StreetViewPanorama;
  getTilt(): number;
  setHeading(heading: number): void;
  setMapTypeId(mapTypeId: google.maps.MapTypeId | string): void;
  setStreetView(panorama: google.maps.StreetViewPanorama | null): void;
  setTilt(tilt: number): void;
  setClickableIcons(clickable: boolean): void;
}

export interface LatLng {
  constructor(lat: number, lng: number): void;
  lat(): number;
  lng(): number;
  toJSON(): any;
  toString(): string;
}
// ...
export interface CircleOptions {
  center?: LatLng | LatLngLiteral;
  clickable?: boolean;
  draggable?: boolean;
  editable?: boolean;
  fillColor?: string;
  fillOpacity?: number;
  map?: GoogleMap;
  radius?: number;
  strokeColor?: string;
  strokeOpacity?: number;
  strokePosition?: 'CENTER' | 'INSIDE' | 'OUTSIDE';
  strokeWeight?: number;
  visible?: boolean;
  zIndex?: number;
}
// ...
export interface RectangleOptions {
  bounds?: LatLngBounds | LatLngBoundsLiteral;
  clickable?: boolean;
  draggable?: boolean;
  editable?: boolean;
  fillColor?: string;
  fillOpacity?: number;
  map?: GoogleMap;
  strokeColor?: string;
  strokeOpacity?: number;
  strokePosition?: 'CENTER' | 'INSIDE' | 'OUTSIDE';
  strokeWeight?: number;
  visible?: boolean;
  zIndex?: number;
}

export interface LatLngBounds {
  contains(latLng: LatLng): boolean;
  equals(other: LatLngBounds | LatLngBoundsLiteral): boolean;
  extend(point: LatLng | LatLngLiteral): void;
  getCenter(): LatLng;
  getNorthEast(): LatLng;
  getSouthWest(): LatLng;
  intersects(other: LatLngBounds | LatLngBoundsLiteral): boolean;
  isEmpty(): boolean;
  toJSON(): LatLngBoundsLiteral;
  toSpan(): LatLng;
  toString(): string;
  toUrlValue(precision?: number): string;
  union(other: LatLngBounds | LatLngBoundsLiteral): LatLngBounds;
}

export interface Padding {
  top: number;
  left: number;
  right: number;
  bottom: number;
}

export interface LatLngBoundsLiteral {
  east: number;
  north: number;
  south: number;
  west: number;
}

export interface LatLngLiteral {
  lat: number;
  lng: number;
}

export interface MouseEvent { latLng: LatLng; }

export interface MapOptions {
  center?: LatLng | LatLngLiteral;
  zoom?: number;
  minZoom?: number;
  maxZoom?: number;
  controlSize?: number;
  disableDoubleClickZoom?: boolean;
  disableDefaultUI?: boolean;
  scrollwheel?: boolean;
  backgroundColor?: string;
  draggable?: boolean;
  draggableCursor?: string;
  draggingCursor?: string;
  keyboardShortcuts?: boolean;
  styles?: MapTypeStyle[];
  zoomControl?: boolean;
  zoomControlOptions?: ZoomControlOptions;
  streetViewControl?: boolean;
  streetViewControlOptions?: StreetViewControlOptions;
  scaleControl?: boolean;
  scaleControlOptions?: ScaleControlOptions;
  mapTypeControl?: boolean;
  mapTypeControlOptions?: MapTypeControlOptions;
  panControl?: boolean;
  panControlOptions?: PanControlOptions;
  rotateControl?: boolean;
  rotateControlOptions?: RotateControlOptions;
  fullscreenControl?: boolean;
  fullscreenControlOptions?: FullscreenControlOptions;
  mapTypeId?: string | MapTypeId;
  clickableIcons?: boolean;
  gestureHandling?: 'cooperative' | 'greedy' | 'none' | 'auto';
  tilt?: number;
  restriction?: MapRestriction;
}

export interface MapTypeStyle {
  elementType?: 'all' | 'geometry' | 'geometry.fill' | 'geometry.stroke' | 'labels' | 'labels.icon' |
  'labels.text' | 'labels.text.fill' | 'labels.text.stroke';
  featureType?: 'administrative' | 'administrative.country' | 'administrative.land_parcel' |
  'administrative.locality' | 'administrative.neighborhood' | 'administrative.province' | 'all' |
  'landscape' | 'landscape.man_made' | 'landscape.natural' | 'landscape.natural.landcover' |
  'landscape.natural.terrain' | 'poi' | 'poi.attraction' | 'poi.business' | 'poi.government' |
  'poi.medical' | 'poi.park' | 'poi.place_of_worship' | 'poi.school' | 'poi.sports_complex' | 'road' |
  'road.arterial' | 'road.highway' | 'road.highway.controlled_access' | 'road.local' | 'transit' |
  'transit.line' | 'transit.station' | 'transit.station.airport' | 'transit.station.bus' |
  'transit.station.rail' | 'water';
  stylers: MapTypeStyler[];
}

/**
 *  If more than one key is specified in a single MapTypeStyler, all but one will be ignored.
 */
export interface MapTypeStyler {
  color?: string;
  gamma?: number;
  hue?: string;
  invert_lightness?: boolean;
  lightness?: number;
  saturation?: number;
  visibility?: string;
  weight?: number;
}
// ...
export interface MapsEventListener { remove(): void; }
// ...
export interface TransitLayerOptions {
  visible: boolean;
}

export interface BicyclingLayer extends MVCObject {
  getMap(): GoogleMap;
  setMap(map: GoogleMap): void;
}

/**
 * Custom type
 */
export interface BicyclingLayerOptions {
  visible: boolean;
}

export interface Data extends MVCObject {
  features: Feature[];
  constructor(options?: DataOptions): void;
  addGeoJson(geoJson: Object, options?: GeoJsonOptions): Feature[];
  remove(feature: Feature): void;
  setControlPosition(controlPosition: google.maps.ControlPosition): void;
  setControls(controls: string[]): void;
  setDrawingMode(drawingMode: string): void;
  setMap(map: GoogleMap): void;
  /* tslint:disable */
  /*
  * Tslint configuration check-parameters will prompt errors for these lines of code.
  * https://palantir.github.io/tslint/rules/no-unused-variable/
  */
  setStyle(style: () => void): void;
  forEach(callback: (feature: Feature) => void): void;
  loadGeoJson(url: string, options?: GeoJsonOptions, callback?: (feats: Feature[]) => void): void;
  /* tslint:enable */
}

export interface Feature extends MVCObject {
  id?: number | string | undefined;
  geometry: Geometry;
  properties: any;
}

export interface SearchBoxOptions {
  bounds?: LatLngBounds|LatLngBoundsLiteral;
}

export interface DataOptions {
  controlPosition?: google.maps.ControlPosition;
  controls?: string[];
  drawingMode?: string;
  featureFactory?: (geometry: Geometry) => Feature;
  map?: GoogleMap;
  style?: () => void;
}

export interface DataMouseEvent extends MouseEvent {
  feature: Feature;
}

export interface GeoJsonOptions {
  idPropertyName: string;
}

export interface Geometry {
  type: string;
}

export enum MapTypeId {
  /** This map type displays a transparent layer of major streets on satellite images. */
  HYBRID,
  /** This map type displays a normal street map. */
  ROADMAP,
  /** This map type displays satellite images. */
  SATELLITE,
  /** This map type displays maps with physical features such as terrain and vegetation. */
  TERRAIN,
}

/***** Controls *****/
/** Options for the rendering of the map type control. */
export interface MapTypeControlOptions {
  /** IDs of map types to show in the control. */
  mapTypeIds?: (MapTypeId | string)[];
  /**
   * Position id. Used to specify the position of the control on the map.
   * The default position is TOP_RIGHT.
   */
  position?: google.maps.ControlPosition;
  /** Style id. Used to select what style of map type control to display. */
  style?: MapTypeControlStyle;
}

export enum MapTypeControlStyle {
  DEFAULT = 0,
  DROPDOWN_MENU = 2,
  HORIZONTAL_BAR = 1,
}

export interface OverviewMapControlOptions {
  opened?: boolean;
}

/** Options for the rendering of the pan control. */
export interface PanControlOptions {
  /**
   * Position id. Used to specify the position of the control on the map.
   * The default position is TOP_LEFT.
   */
  position?: google.maps.ControlPosition;
}

/** Options for the rendering of the rotate control. */
export interface RotateControlOptions {
  /**
   * Position id. Used to specify the position of the control on the map.
   * The default position is TOP_LEFT.
   */
  position?: google.maps.ControlPosition;
}

/** Options for the rendering of the scale control. */
export interface ScaleControlOptions {
  /** Style id. Used to select what style of scale control to display. */
  style?: ScaleControlStyle;
}

export enum ScaleControlStyle {
  DEFAULT,
}

/** Options for the rendering of the Street View pegman control on the map. */
export interface StreetViewControlOptions {
  /**
   * Position id. Used to specify the position of the control on the map. The
   * default position is embedded within the navigation (zoom and pan) controls.
   * If this position is empty or the same as that specified in the
   * zoomControlOptions or panControlOptions, the Street View control will be
   * displayed as part of the navigation controls. Otherwise, it will be displayed
   * separately.
   */
  position?: google.maps.ControlPosition;
}

/** Options for the rendering of the zoom control. */
export interface ZoomControlOptions {
  /**
   * Position id. Used to specify the position of the control on the map.
   * The default position is TOP_LEFT.
   */
  position?: google.maps.ControlPosition;
  style?: ZoomControlStyle;
}

export enum ZoomControlStyle {
  DEFAULT,
  LARGE,
  SMALL,
}

/** Options for the rendering of the fullscreen control. */
export interface FullscreenControlOptions {
  /**
   * Position id. Used to specify the position of the control on the map.
   * The default position is RIGHT_TOP.
   */
  position?: google.maps.ControlPosition;
}

/** Options for the restricting the bounds of the map. */
export interface MapRestriction {
  latLngBounds: LatLngBounds | LatLngBoundsLiteral;
  strictBounds?: boolean;
}